#176 feat: Add notification toggle buttons in portfolio page
