+++ src/components/Portfolio/PortfolioOverview/PortfolioOverviewBody.tsx
@@ -1,0 +1,257 @@
+import usePortfolioNotificationSettingsMutation from "@api/notifications/queries/usePortfolioNotificationSettingsMutation";
+import { PortfolioDetails, PortfolioDetailsSSE } from "@api/portfolio/types";
+import RateBadge from "@components/common/Badges/RateBadge";
+import { CustomTooltip } from "@components/common/CustomTooltip";
+import { Icon } from "@components/common/Icon";
+import { IconButton, debounce } from "@mui/material";
+import designSystem from "@styles/designSystem";
+import { thousandsDelimiter } from "@utils/delimiters";
+import { useParams } from "react-router-dom";
+import styled from "styled-components";
+
+type Props = {
+  data: PortfolioDetails;
+  sseData: PortfolioDetailsSSE | null;
+};
+
+export default function PortfolioOverviewBody({ data, sseData }: Props) {
+  const { portfolioId } = useParams();
+
+  const { mutate } = usePortfolioNotificationSettingsMutation(
+    Number(portfolioId)
+  );
+
+  const onTargetGainNotifyButtonClick = debounce(() => {
+    mutate({
+      notificationType: "targetGain",
+      body: { isActive: !data.targetGainNotify },
+    });
+  }, 250);
+
+  const onMaxLossNotifyButtonClick = debounce(() => {
+    mutate({
+      notificationType: "maxLoss",
+      body: { isActive: !data.maxLossNotify },
+    });
+  }, 250);
+
+  return (
+    <StyledOverviewBody>
+      <OverviewBodyTop>
+        <OverviewBodySection>
+          <OverviewBodyData>
+            <div>예산</div>
+            <span>₩{thousandsDelimiter(data.budget ?? 0)}</span>
+          </OverviewBodyData>
+          <OverviewBodyData>
+            <div>투자금액</div>
+            <span>₩{thousandsDelimiter(data.investedAmount ?? 0)}</span>
+          </OverviewBodyData>
+          <OverviewBodyData>
+            <div>잔고</div>
+            <span>₩{thousandsDelimiter(data.balance ?? 0)}</span>
+          </OverviewBodyData>
+          <OverviewBodyData>
+            <CustomTooltip
+              arrow
+              placement="bottom-start"
+              title={
+                <p>
+                  손실 종목을 매도 시 현재 잔고의 감당력을 표현하는 것으로 매도
+                  후 실제 잔고를 나타내는 것이 아닙니다
+                  <br />
+                  잔고 - 손실 종목의 손실 합
+                </p>
+              }>
+              <div
+                style={{ display: "flex", alignItems: "center", gap: "4px" }}>
+                잠정 손실 잔고 <Icon icon="help" size={16} color="gray400" />
+              </div>
+            </CustomTooltip>
+            <span>
+              ₩
+              {thousandsDelimiter(
+                sseData?.provisionalLossBalance ?? data.provisionalLossBalance
+              )}
+            </span>
+          </OverviewBodyData>
+        </OverviewBodySection>
+        <OverviewBodySection>
+          <OverviewBodyData>
+            <NotificationLabel>
+              목표 수익률{" "}
+              <NotificationIconButton onClick={onTargetGainNotifyButtonClick}>
+                <Icon
+                  icon="notification"
+                  size={24}
+                  color={data.targetGainNotify ? "blue500" : "gray400"}
+                />
+              </NotificationIconButton>
+            </NotificationLabel>
+            <span>₩{thousandsDelimiter(data.targetGain ?? 0)}</span>
+          </OverviewBodyData>
+          <div style={{ marginLeft: "auto" }}>
+            <RateBadge
+              size={16}
+              value={data.targetReturnRate}
+              bgColorStatus={false}
+              iconStatus={false}
+            />
+          </div>
+          <OverviewBodyData>
+            <NotificationLabel>
+              최대 손실율{" "}
+              <NotificationIconButton
+                onClick={onMaxLossNotifyButtonClick}
+                sx={{
+                  height: "24px",
+                  padding: "0",
+                }}>
+                <Icon
+                  icon="notification"
+                  size={24}
+                  color={data.maxLossNotify ? "blue500" : "gray400"}
+                />
+              </NotificationIconButton>
+            </NotificationLabel>
+            <span>₩{thousandsDelimiter(data.maximumLoss ?? 0)}</span>
+          </OverviewBodyData>
+          <div style={{ marginLeft: "auto" }}>
+            <RateBadge
+              size={16}
+              value={-data.maximumLossRate}
+              bgColorStatus={false}
+              iconStatus={false}
+            />
+          </div>
+        </OverviewBodySection>
+      </OverviewBodyTop>
+
+      <OverviewBodyBottom>
+        <OverviewBodySection>
+          <OverviewBodyData>
+            <div>총 손익</div>
+            <span>
+              ₩{thousandsDelimiter(sseData?.totalGain ?? data.totalGain)}
+            </span>
+          </OverviewBodyData>
+          <div style={{ marginLeft: "auto" }}>
+            <RateBadge
+              size={16}
+              value={sseData?.totalGainRate ?? data.totalGainRate}
+              bgColorStatus={false}
+              iconStatus={false}
+            />
+          </div>
+          <OverviewBodyData>
+            <div>당일 손익</div>
+            <span>
+              ₩{thousandsDelimiter(sseData?.dailyGain ?? data.dailyGain)}
+            </span>
+          </OverviewBodyData>
+          <div style={{ marginLeft: "auto" }}>
+            <RateBadge
+              size={16}
+              value={sseData?.dailyGainRate ?? data.dailyGainRate}
+              bgColorStatus={false}
+              iconStatus={false}
+            />
+          </div>
+        </OverviewBodySection>
+        <OverviewBodySection>
+          <OverviewBodyData>
+            <div>총 연배당금</div>
+            <span>₩{thousandsDelimiter(data.annualDividend)}</span>
+          </OverviewBodyData>
+          <div style={{ marginLeft: "auto" }}>
+            <RateBadge
+              size={16}
+              value={data.annualDividendYield}
+              bgColorStatus={false}
+              iconStatus={false}
+            />
+          </div>
+          <OverviewBodyData>
+            <CustomTooltip
+              arrow
+              placement="bottom-start"
+              title={
+                <p style={{ font: designSystem.font.body4.font }}>
+                  총 투자금액 대비 연배당률
+                  <br />연 배당금 / 투자금액 * 100%
+                </p>
+              }>
+              <div
+                style={{ display: "flex", alignItems: "center", gap: "4px" }}>
+                투자대비 연 배당률
+                <Icon icon="help" size={16} color="gray400" />
+              </div>
+            </CustomTooltip>
+            <RateBadge
+              size={16}
+              value={data.annualInvestmentDividendYield}
+              bgColorStatus={false}
+              iconStatus={false}
+            />
+          </OverviewBodyData>
+        </OverviewBodySection>
+      </OverviewBodyBottom>
+    </StyledOverviewBody>
+  );
+}
+
+const StyledOverviewBody = styled.div`
+  display: flex;
+  flex-direction: column;
+  border: 1px solid ${designSystem.color.neutral.gray100};
+  border-radius: 8px;
+  font: ${designSystem.font.title5.font};
+  letter-spacing: ${designSystem.font.title5.letterSpacing};
+  color: ${designSystem.color.neutral.gray600};
+  overflow: hidden;
+`;
+
+const OverviewBodyTop = styled.div`
+  display: flex;
+  border-bottom: 1px solid ${designSystem.color.neutral.gray100};
+`;
+
+const OverviewBodyBottom = styled.div`
+  display: flex;
+`;
+
+const OverviewBodySection = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+  width: 444px;
+  height: 140px;
+  padding: 16px;
+
+  &:first-child {
+    border-right: 1px solid ${designSystem.color.neutral.gray100};
+  }
+`;
+
+const OverviewBodyData = styled.div`
+  height: 24px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+
+  > span {
+    font: ${designSystem.font.body3.font};
+    color: ${designSystem.color.neutral.gray900};
+  }
+`;
+
+const NotificationLabel = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 4px;
+`;
+
+const NotificationIconButton = styled(IconButton)`
+  height: 24px;
+  padding: 0;
+`;
